PP-4512 Add test for case where epdq status is unknown and fix
ePDQ can respond with a parseable respnse with an unknown status.
Rather than pass nulls up the stack, I have made the status mapper
throw an error, which at the moment is caught and dealt with correctly by ChargeExpiryService. The error handling and passing can be
improved, but this commit ensures Expiry sweep works in all cases.