PP-4560 Make JsonPatch.valueAs… methods not return null
Make JsonPatch.valueAsString() and JsonPatch.valueAsLong() throw
exceptions rather than return null if the node cannot be converted
to the appropriate type.

Remove JsonNode.valueAsList() because it’s unused.

This change is safe to make because the validation in
GatewayAccountRequestValidator (which everything that is used to
construct a JsonPatchRequest goes through) ensures that the nodes
of appropriate types as described below.

valueAsString():
• "credentials/gateway_merchant_id" — checks not null
• "allow_google_pay" — checks "true" or "false"
• "allow_apple_pay" — checks "true" or "false"
• "allow_web_payments" — checks "true" or "false"
• "email_collection_mode" — checks if EmailCollectionMode variant

valueAsLong():
• "corporate_credit_card_surcharge_amount" — checks +ve number
• "corporate_debit_card_surcharge_amount" — checks +ve number
• "corporate_prepaid_credit_card_surcharge_amount" — checks +ve num
• "corporate_prepaid_debit_card_surcharge_amount" — checks +ve num

with @stephencdaly