PP-4882 Refactor GatewayAccountDao to not return GatewayAccountResourceDTO
GatewayAccountDao is a JPA dao paramaterized to type GatewayAccountEntity. This
changes the `list` and `listAll` methods to return `GatewayAccountEntity` and
remove the concern of converting `GatewayAccountEntity` to a `GatewayAccountResponseDTO`.
This is in the interest of maintaining a sensible seperation of concerns, to
permit future refactoring of the numerous Gateway Account return type objects and move
towards a common means of fetching Gateway Accounts from the database and converting
to the required return type object (at the time of this change there are numerous ways
which causes confusion to new team members, especially when adding a new attribute
to Gateway Account.
- Change `list` and `listAll` methods to return a list of `GatewayAccountEntity`
- Add `.fromEntity` static method to `GatewayAccountResourceDTO`.
- Move resonsbility for converting between the entity and DTO into the
`GatewayAccountService`.
- Update test and add new test for `.fromEntity` method.