PP-4856 WorldpayStatus and WorldPayQueryResponse
This commit adds WorldPayQueryResponse to deserialise responses
to Worldpay status queries. This is based on our other templates,
but I have removed any interpretative code - it should just
convert the XML to a very basic POJO.
The second things introduced is a WorldpayStatus enum. This
can be used to interpret the WorldPayQueryResponse robustly -
it maps the query response to a known Pay status, or returns
an empty Optional.
I have included tests to show this, using the example query
responses provided by Worldpay
http://support.worldpay.com/support/kb/gg/corporate-gateway-guide/content/manage/inquiryrequests.htm