PP-4898 Remove withoutCredentials method in favour of GatewayAccountResourceDTO
Remove the use of `GatewayAccountEntity.withoutCredentials()` in favour of the
existing `GatewayAccountResourceDTO` to rationalise the number of return type
objects and processes surrounding gateway accounts in Connector.
- Add "email_notifications", "email_collection_mode" and "toggle_3ds" to `GatewayAccountResourceDTO`
so that this Type returns the same information as the `Map` produced by `withoutCredentials`.
- This changes the type of the values contained within the response since `withoutCredentials`
was converting each type to String instead of using boolean for example.
- Update tests to assert on type boolean rather than expecting the String equivelant
e.g. `"true"'.
- Update CardResource to use `GatewayAccountResourceDTO.fromEntity()` in place
of calling `withoutCredentials()`.