PP-4877 Increase capture retry count default value
Increase to 96 as we had some issues with EPDQ not sending the captured
notification for over a day.

2 days should be sufficient for this, i.e.
CAPTURE_PROCESS_RETRY_FAILURES_EVERY set to 60 and
CAPTURE_PROCESS_MAXIMUM_RETRIES set to 48

However due to how the current capture process works, it is possible for
more than one connector node to attempt to capture the same charge in an
hour, so the number of retries does not equal the number of hours since
the first capture attempt.

This change has been made as a temporary solution as we will be looking
at changing how the capture process works in the near future.

with @heathd and @cobainc0