PP-4896: Delete superfluous AsymmetricKeyVerifier
The AsymmetricKeyVerifier was creating a digital signature using a randomly
generated bytes and the private key, and then verifying
itself, which is pretty useless. If the public and private keys don't match,
decryption would fail anyway. Worse still it's `verify` method was taking 6
minutes, most likely due to a lack of entropy on the AWS environment while
generating the random bytes. This can take a few minutes.