PP-4758: Differentiate between 4xx and 5xx exception
Callers of StripeGatewayClient handle 4xx and 5xx differently by virtue of
StripeGatewayClient throwing separate exceptions for each. This commit
introduces GatewayErrorException.DownstreamErrorException (for 4xx status
codes) and GatewayErrorException.ClientErrorException (for 5xx status codes) to the GatewayClient to
make deletion of StripeGatewayClient easier later. It's also probably a
good idea so we can tell whether it's connector making a bad request or a
downstream error from a payment provider.