PP-4994: Use Stripe charge `transfer_data` API
Stripe's charge creation API has been updated with no direct mention of
the `destination[amount]` syntax that we use for creating charges.

Update the params to use the `transfer_data[destination]` syntax and
explicitly set `on_behalf_of` to specify the settlement merchant to use
for the transaction, this is no longer implicit.

The responses from Stripe for contractual tests may need to be updated
to reflect the new attributes that are now set.