PP-4993 Introduce Stripe*Request objects
Adds StripeRefundRequest and StripeCaptureRequest.
This may seem like overkill, but makes it easier to reason
about communications with Stripe, and encapsulates some of the
logic which previously needed to be done by hand in several
places, such as adding the Authorization header.

As part of this I have created a GatewayClientRequest interface
and a method to use this in GatewayClient, which makes
calling GatewayClient quite seamless.