Introduce Nettable interface
In order to share net amount calaculation logic between
ChargeEntity and Transaction, this commit introduces
the Nettable interface which has a default implementation
of `getNetAmount`.
In future it might be nice to refactor this to have
methods like `getDeductions` and `getAdditions`, which
would be more generic, but I think this is fine for
now.