PP-5142 Use ErrorResponse to construct responses created by ResponseUtil
Modify ResponseUtil so that most of the methods that construct an error
response use the ErrorResponse model.
Modify tests to expect a "messages" array and to assert on the
"error_identifier" - which has a value of "GENERIC" for all responses
currently.
Not all responses have been converted to use ErrorResponse yet.