PP-4993 Authorise stripe charges without setting destination
In order to start taking fees via the method of doing a net transfer
from our Stripe platform account to the relevant Stripe connect
account, we need to stop telling Stripe to automatically transfer
a charge to a connect account on capture. This commit does that.

Whilst this change is small, it will have a quite a big effect, as
it will activate the code path here
https://github.com/alphagov/pay-connector/blob/master/src/main/java/uk/gov/pay/connector/gateway/stripe/handler/StripeCaptureHandler.java#L50

which will mean captures will now be processed by a two stage process.
Therefore, to test this PR I recommend testing locally against real Stripe
gateway and ensuring that authorisation, capture, and refund all behave as expected.