PP-5237: Expire charges in auth error state
When making 3ds Authorisation requests, the EPDQ server sometimes responds with
a 503 status, which causes the charge to be transitioned to AUTHORISATION ERROR
in connector. It appears that despite the 503, the charge will still be in the
AUTHORISED SUCCESS (or 5-Authorised) state. This makes sense as it was already
authorised in the first authorisation request before the 3ds one. This commit
therefore adds the AUTHORISED ERROR state as one of the statuses upon which the
charge should be cancelled.

I had considered using the DiscrepancyService to determine if there is a
discrepancy before cancelling the charge; however looking at EpdqStatusMapper
and WorldpayStatus classes these PSPs never have a state that is equivalent to
our AUTHORISATION ERROR status so I surmise it is always safe to call cancel on
a charge that has errored.

Add test using epdq as a PSP to spice things up.