zd3689082: Use the MDC to set the chargeId
This commit uses an implementation of a
javax.ws.rs.container.ContainerRequestFilter to put the "chargeId" value into
the MDC. I used a ContainerRequestFilter rather than a javax.servlet.Filter (see
https://github.com/alphagov/pay-publicapi/blob/master/src/main/java/uk/gov/pay/api/filter/RateLimiterFilter.java
for an example) because its filter method has a ContainerRequestContext has a
nice API for getting the chargeId; so in the ChargeIdLoggingMDCRequestFilter:

```
requestContext.getUriInfo().getPathParameters().getFirst("chargeId")
```

The "chargeId" string is defined in the `@Path` annotations in the CardResource
(`@Path("/v1/frontend/charges/{chargeId}/3ds")` for example).

The obvious benefit of this is that one doesn't need to parse the URI, as that
is would one would have to do if an implementation of javax.servlet.Filter was
used.

Another benefit is that we don't have to set the chargeId in the MDC manually
in each resource method and use a try...finally block to remove the chargeId
from the MDC. That would be very ugly in my opinion as resource code shouldn't
concern itself with MDC stuff.

The downside is that we need a ChargeIdLoggingMDCResponseFilter to remove the
chargeId from the MDC.

I followed
https://www.dropwizard.io/1.3.11/docs/manual/core.html#jersey-filters to use a
ContainerRequestFilter in dropwizard.

Requests to CardResource usually spin off another thread to perform an
operation though. The values in the original MDC do not get propagated into
these new threads. I therefore had to modify the CardExecutorService to inherit
the MDC context from the originating thread.

Unfortunately, testing these things is impossible in my opinion. You can
satisfy yourself that this works by running
https://github.com/alphagov/pay-connector/blob/d74c53b539489cf0aa5a9a4542aacc940f8681ae/src/test/java/uk/gov/pay/connector/it/resources/worldpay/WorldpayCardResourceIT.java#L273
in Intellij. You should see something like

```
[2019-05-23 16:01:53.379] [thread=CardExecutorService-0] [level=ERROR] [logger=u.g.p.c.g.GatewayClient] [requestID=(none)] [chargeID=charge-1186667931] - Exception for gateway url=http://localhost:57596/jsp/merchant/xml/paymentService.jsp, error message: java.net.SocketException: Connection reset
```

 logged.