BAU Ensure stable ordering in getInternalEvents
There are tests that check the order of internal events

however this SQL query did not specify an explicit ordering on events. Without
an explicit `ORDER BY` clause there is no guarantee of the order of events
returned by the query, so this test could be flaky (and has been observed to be
flaky at least in local development environments) due ot the unstable ordering.

This PR adds an explicit stable ordering to the test helper.