PP-5409 Improve cancellation with gateways
Use same methodology as charge expiry to determine whether charges need
cancelling with the gateway. This groups charge statuses into
PRE_AUTHORISATION, DURING_AUTHORISATION and POST_AUTHORISATION. For
charges with statuses in the DURING_AUTHORISATION category, we check
with the gateway to determine whether we need to send a cancellation
request.

Also add the ability to cancel charges in AUTHORISATION_READY and
AUTHORISATION_3DS_READY statuses, as we should be able to cancel a
charge at any point when it isn't in a terminal state.