PP-5386 Undefined charge status
Introduces a new undefined charge status, and on
charge creation initialises a charge as undefined,
then transitions it to created. This means a PaymentCreated event is emitted in the same way all
other events are emitted.
The charge entity needs to be merged after transitioning
to created, because for some reason that I haven't
got totally to the bottom of, after the call to persist
the charge entity is detached. I have chosen to do
the merge as a one off thing, rather than move it in
to the transitionChargeStatus method, as it is
seemingly only needed after entity creation.