BAU don't pact test against everything
when running provider tests we are provided with an environment variable `PACT_CONSUMER_TAG`.

When running a branch or master build of the provider this will be set to
`master`. So in this scenario we would test the provider agasint the master
build of the consumer.

When running a provider test in the context of a consumer build via:
  https://github.com/alphagov/pay-jenkins-library/blob/master/vars/runProviderContractTests.groovy

eg. https://github.com/alphagov/pay-publicapi/blob/master/Jenkinsfile#L98-L116

the `PACT_CONSUMER_TAG` will be set to the git sha of the consumer. The
consumer will have run and published its pacts and tagged them immediately
before this step.

In neither of these circumstances is it useful to have the provider also check
against the `test`, `staging` or `production` versions of the pacts.

Having these additional checks makes it harder (or maybe impossible?) to evolve
contracts between services.

Those checks can happen upon deployment eg.

https://github.com/alphagov/pay-connector/blob/master/Jenkinsfile#L144