PP-5391 Emit refund events
Emit refund state transition events by hooking into
centralised `transitionRefundStatus` method. We re-use
the approach taken for emitting payment state
transitions, whereby the transitions is put on a queue
and then read off and processed, to ensure we only
emit events that completed successfully, and to be
able to use DB based timestamp.