PP-5463 Refund availability event
In order to keep track of the amount available to be refunded
in ledger, we need to emit `RefundAvailabilityUpdated` eevnts from
connector whenever refundability of a payment might change.
To do this I have changed how refund state transtions are
handled, so that rather than triggering the emsission of a single
external event it can emit a sequence of events (currently only
a sequence of either 1 or 2). The logic for all this has been moved
to the RefundEventFactory.
The refund availability of a payment only changes on refund creation
when the amount available to be refunded goes down) and refund error
(when it goes back up). The actual payload of the events is calculated
using the same tools that are currently used to populate responses,
so everything will line up.