PP-5520: delay state transition process w/ blocking queue
State transition events are currently rate limited by a scheduled
exectutor service waiting delaying polling by 100ms. This ensures
Connector is not constantly polling but doesn't allow the in-memory
queue to be cleared effectively given many state transitions.

* Use the blocking queue poll timeout interface to wait for a `Delayed`
entry to become visible -- this includes the delay if there is nothing
to process but lets the emitter process all elements if there is backlog
* Just use an executor service as there is no scheduling required