PP-5547 Add Json Web Token support
Create a JWTGenerator class to offer simple creation of JWTs. This uses
`io.jsonwebtoken` which is listed as a reputable library on `jwt.io`, it
has a simple builder pattern to generate the tokens and an Apach 2.0
licence. It enforces key strength when signing and will not permit the
use of a key considered too weak for the desired algorithm (e.g. for
HSA256 the key must be greater than 256 bits). Since the Worldpay test
token is only 288 bits and we cannot change it, this initial
implementation uses HSA256 to sign the JWT, which is supported by
Worldpay.

This commit includes the first public method on the generator to generate a
Worldpay 3DS Flex DDC token and a unit test to ensure the token, when
decoded includes the correct header and body elements.