bau: Log when dropwizard shutdowns in tests
Am not sure why SendRefundEmailIT.shouldSendEmailFollowingASuccessfulRefund
occasionally fails. When it does there is a log `Email notifications is
disabled by configuration` which would explain why the test failed but at this
point I'm not sure why dropwizard was started with a wrong config, as
SendRefundEmailIT overrides the `notifyConfig.emailNotifyEnabled` to be `true`.