PP-5545: Don't supply authenticationRiskData element
This is change in acceptance criteria. We're not going to supply this anymore
as we do not authenticate paying users and none of the options for the
`authenticationMethod` attribute of the `authenticationRiskData` elements
(guestCheckout, localAccount, federatedAccount, fidoAuthenticator,
issuerCredentials, thirdPartyAuthentication) work for us. This element is also
specified as optional in the worldpay docs.