PP-5610 Remove validation annotations form Address
There is a new 'PrefilledAddress' class used to for when a new charge is
created and has a 'PrefilledCardHolderDetails'. The original 'Address' class is now
used when Frontend updates the card details and for this end-point the
javax.validation annotations are not yet being used. This commit removes
the annotations to avoid confusion.