PP-5613 Enforce ExternalMetadata Validation when saving ChargeEntity
We evaluate the `@ValidExternalMetadata` when accepting a create charge
request via an end-point however this is not evaluated when attempting
to persist a `ChargeEntity`. This means it is possible to create an
invalid instance of `ExternalMetadata` within Connector and then persist
it to the charges table. This change adds the `@Valid` annotation to the
`externalMetadata` attribute of `ChargeEntity` so that any JPA actions
to update or persist the entity will evaluate the valiation, rolling
back the transaction if it fails.