PP-5585 Stripe capture with payment intents
We need to switch over to using Stripe payment intents to process
payments with Stripe, in order to be 3DS2 compliant. In order to
do this in a backwards compatible way we need the capturing with
stripe to work with charges created both via the charges api and
the payment intents api.
This PR uses the fact that stripe namespaces all their ids with
a prefix indicating the resource type (so a payment intent always
has a prefix 'pi_...') so that we can capture the charge
appropriately.
This PR should not affect existing payment flow at all, and will
need to be tested more thoroughly once changes to authorisation
have been implemented.