PP-5619 Add `used` to `TokenEntity`
`used` is set to `false` when creating new tokens. Update tests to
check.

We use a primitive boolean for the field in the `TokenEntity`. When JPA loads
an existing token from the databse with a null value in the `used` column, this
will get mapped to `false` (we have tested this), which is what we want.

with @AlexBishop1