PP-5585 Robustify authorisation with payment intents
There is a non-zero probability that Stripe will respond to a
payment intent creation request with a status of `requires_confirmation`.
In this case we need to go back to Stripe to confirm.

This commit changes Stripe authorisation handler to handle this case. It
assumes that Stripe will not respond to a confirmation
request with the status as `requires_confirmation` again (otherwise I have
created an infinite loop...).