PP-5672 Make Stripe refund more robust
This commit mitigates a problem we have observed in the wild.
The problem is that the following can happen:
- refund request to Stripe succeeds
- transfer request from connected account to platform account fails
- we as GOV.UK Pay are left out of pocket
- from service's point of view refund seems to be an error, yet the
end user has recieved the refund, and the service still recieves the
funds as part of next payout (since it is still credited to their account)

To avoid this, the `StripeRefundHandler` now does the following:
- transfer funds from connect account to platform account (if this
fails, refund fails, and no harm done)
- attempt to refund payment
- if refund step fails, reverse original transfer
- iff reversal fails, log error that will be picked up by Splunk alerting

The general idea is that we try reversable things before non-reversable things,
so that in the case of failure of non-reversible operation we can revert the
whole state without causing confusion.

I have also tidied up the tests somewhat. There were tests claiming to be about
testing specific kinds of failure from Stripe, which were essentially duplicate
since all Stripe failures are being mocked. I have left tests that check the
behaviour of refund handler for main failure modes that the handler can
discriminate between.

Finally, I have removed handling of 'non-payment intents' based payments
as we now only process payments using payment intents API.