PP-5696 Log Smartpay 3D secure response
Smartpay 3D secure is currently broken, we don't seem to be extracting
the response correctly so the payment is going into AUTHORISATION ERROR
state when it has been successfully AUTHORISED in Smartpay.

Log the entire response so we can figure out what we're getting back
when the Authorisation is successful.