PP-5204 Handle low value Stripe payments
For reasons too complicated to explain we capture Stripe payments
by first capturing the full amount to the Pay platform account, then
transferring the balance (total - fee) to the connect account.
In the case where fee > amount, we actually need to recoup the
(fee - amount) from the connect accoun in order to break even.
If fee = amount we do nothing.