PP-5704: update column not to be insertable or updatable by persistance provider
Currently the connector nodes generate timestamps that are later put in the db.
This can sometimes cause the problem, when the nodes have different values of time.
In order to mitigate the problem we're going to let db set the timestamp values for
ChargeEvents.
The way we currently handle timestamps is causing some problems in the local environment (if
the timezone is different than UTC).
The long term solution should be to convert all timestamps (at the moment all of them are without
a timzone info) to timestamps with timezone.
The fix put in this PR allows connector to read the timestamp data as UTC regardless the timezone
the app runs in (the explaination can be found here:
https://stackoverflow.com/questions/41618165/how-to-specify-utc-timezone-for-spring-boot-jpa-timestamp).

Changes (tl;dr):
* ChargeEventEntity.updated column changed to non-insertable and non-updatable (db sets the value)
  * replaced its converter to be able to read the value as UTC regardless the jvm timezone
* update countChargesForImmediateCapture (used for metrics only) dao method to pass the cutoff date in UTC
* update DatabaseHelper.addEvent to insert timestamp in UTC