PP-5666: add pre-filled card details to payment created event
Why?
To allow searches on pre-filled data in ledger

Why is it complicated?
* it’s impossible to distinguish the new event from event backfilling
* once user fills the card details the pre-filled data is lost

Why does this approach work?
During the backfill the states happening after AUTHORISATION_READY event
(AUTHORISATION_READY won’t appear in db) will cause PaymentDetailsEntered
to be emitted. So if there is a charge (in whatever state) that doesn’t
have any charge events in the post-AUTHORISATION_READY states
(PaymentGatewayStateTransitions.java#L112-L120) then if we have information
about card details we can freely add it to PaymentCreated event.
In the normal event flow (not a backfill) this conditions should apply too
(there shouldn’t be enough time to transition to the next state (any of the
post-auth-ready ones) and if there is then we’ve lost the pre-filled data
anyway so we shouldn’t add them to the PaymentCreated payload.

Changes:
* updated PaymentDetailsEntered event to include card-details and email
* added relevant unit tests