PP-5781: transition from 3DS flow to external cancel request
A number of requests are failing because we do not currently support
transitioning from 3DS REQUIRED to system cancelled.

This has reportedly caused issues in the past with our expiry sweeper
however these issues have been addressed.

It seems to be a valid transition tested on Worldpay accounts.
Determining how sensible this is for all Gateways might be required.